+++ lib/jobs/bard.lua
@@ -596,11 +596,16 @@ return {
     -- Default settings for UI
     default_settings = {
         buff_enabled = true,
+        rest_enabled = false,
+        rest_timer = 5,
+        rest_threshold = 70,
+        rest_distance = 7,
     },
     
     -- Action priority order
     priority_order = {
         'item',
         'buff',
+        'rest',
     }, 
 }