+++ lib/jobs/red_mage.lua
@@ -483,6 +483,10 @@ return {
         focus_enabled = false,
         focus_target_index = nil,
         focus_threshold = 85,
+        rest_enabled = false,
+        rest_timer = 5,
+        rest_threshold = 70,
+        rest_distance = 7,
     },
     
     -- Action priority order
@@ -492,6 +496,7 @@ return {
         'wake',
         'recover',
         'buff',
+        'rest',
         -- revive,
     },
 }